w is double in cube_ht now, also cube's --- prints are commented

+++ cube_ht.h
@@ -25,7 +25,7 @@ public:
   //int size;
   //std::vector< std::vector< std::pair<my_vector<T> * , long int> > > table; //pinakas (o hash table) apo pinakes (h lista ana bucket) apo zeugaria (zeugari dianusmatos kai timhs g)
   //g_funs<T> my_g;
-  cube_ht<T>(int d_ton, int vector_dimensions, int w_to_be); //constructor
+  cube_ht<T>(int d_ton, int vector_dimensions, double w_to_be); //constructor
   ~cube_ht();
   int fi_function(long int gval, int index);
   void cubify_vector(my_vector<T> *v);
